feat!: Teach queries insert, update, delete, which repos now use
This reduces duplication between the query and repository classes, and pushes
the rest of the common backend knowledge down to the query class. Repositories
still need to be backend-specific, as they need to know how to instantiate a
session, and may have the option of doing backend-specific things that shouldn't
be pushed to the query class, but concrete repository implementations can now be
vastly simplified.

BREAKING CHANGE: `query_class` *must* now be defined on a repository subclass.